fix: updateElapsedTime 순수함수 분리
